requirements.txt
pandas
numpy
scikit-learn
xgboost
joblib
requests
beautifulsoup4
faker
lxml
matplotlib
rapidfuzz
tldextract
warcio
dotenv
datasets
huggingface_hub
mysql-connector-python